BroadleafCommerce/QA#2949 - Split the DefaultSEOVariableExpression functionality into the SeoVariableExpression that leverages the SeoPropertyService and the SeoElementVariableExpression which helps to render portions of the Product/Category titles & meta descriptions
